Split WorkersTimeline into composables and presentational sub-components
The 795-line single-file component mixed data transformation, viewport
(zoom/pan) state, SVG geometry, color helpers, and all rendering. Break it into
focused, individually maintainable pieces with no behavior change:

- useTimelineModel — pure transform from test cases to drawable rows/worker
  rows/shard groups/maxTime. Now unit-tested (tests/unit/timeline-model.test.ts)
  to lock the interleaved + sequential-fallback behavior.
- useTimelineViewport — zoom/pan/fit, ResizeObserver, px-per-ms scale, bar
  geometry and axis ticks. Collapses the two duplicated live-fit watchEffects
  into one.
- app/utils/timeline.ts — pure color, time-format and layout-constant helpers.
- timeline/ sub-components: TimelineHeader, TimelineGrid (axis/ticks/zebra/labels/
  shard separators), TimelineBar (the four bar variants), TimelineTooltip.
- WorkersTimeline.vue is now a ~150-line orchestrator that owns the <svg> root
  plus tooltip state and wires the above together.

vitest.config.ts gains Nuxt's path aliases so app composables can be unit-tested.
Verified end to end: the timeline renders with correct SVG namespace, dense
per-worker bars, and wasted-time overlays; typecheck, lint, format and 145 unit
tests pass.
